refactor(hub): split HubDashboard + collapse hero-card duplication
HubDashboard.jsx (661 LOC) -> 140 LOC + three focused files:
- src/core/dashboard/moduleConfigs.jsx   (MODULE_CONFIGS + safeReadJSON)
- src/core/dashboard/DailyProgressHero.jsx (readDailyProgress + hero)
- src/core/dashboard/ModuleCard.jsx       (ModuleCard + SortableCard)

Inline <svg> icons in MODULE_CONFIGS replaced with <Icon /> calls (credit-card,
dumbbell, check, utensils); chevron-right in ModuleCard likewise. Per-module
color+gradient tables extracted from render to module scope.

Card.tsx: finyk/fizruk/routine/nutrition variants now carry their dark-mode
border + subtle tinted overlay, so callers can drop the CSS class equivalents.

index.css: remove .finyk-hero-card / .fizruk-hero-card / .routine-hero-card /
.nutrition-hero-card (and their .dark overrides) - Card variant is the single
source of truth. RoutineCalendarPanel migrated to <Card variant=routine>.
